refactor: package 구조 변경 domain에 관련된 클래스들을 domain 패키지로 이동
