+++ src/main/java/com/codessquad/qna/api/question/ApiAnswerController.java
@@ -3,11 +3,11 @@ package com.codessquad.qna.api.question;
 import com.codessquad.qna.api.result.Result;
 import com.codessquad.qna.common.error.exception.LoginRequiredException;
 import com.codessquad.qna.common.utils.HttpSessionUtils;
-import com.codessquad.qna.web.question.Answer;
-import com.codessquad.qna.web.question.AnswerRepository;
-import com.codessquad.qna.web.question.Question;
-import com.codessquad.qna.web.question.QuestionRepository;
-import com.codessquad.qna.web.user.User;
+import com.codessquad.qna.domain.question.Answer;
+import com.codessquad.qna.domain.question.AnswerRepository;
+import com.codessquad.qna.domain.question.Question;
+import com.codessquad.qna.domain.question.QuestionRepository;
+import com.codessquad.qna.domain.user.User;
 import org.springframework.web.bind.annotation.*;
 
 import javax.servlet.http.HttpSession;
