+++ src/main/java/com/codessquad/qna/web/question/QuestionController.java
@@ -3,7 +3,11 @@ package com.codessquad.qna.web.question;
 import com.codessquad.qna.common.error.exception.LoginRequiredException;
 import com.codessquad.qna.common.error.exception.QuestionNotFoundException;
 import com.codessquad.qna.common.utils.HttpSessionUtils;
-import com.codessquad.qna.web.user.User;
+import com.codessquad.qna.domain.question.Answer;
+import com.codessquad.qna.domain.question.AnswerRepository;
+import com.codessquad.qna.domain.question.Question;
+import com.codessquad.qna.domain.question.QuestionRepository;
+import com.codessquad.qna.domain.user.User;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
